requirements.txt
Django==1.5.5
Fabric==1.8.0
MySQL-python==1.2.5
Pillow==2.2.2
South==0.8.4
argparse==1.2.1
distribute==0.7.3
django-localflavor==1.0
django-simple-captcha==0.4.0
django-tinymce==1.5.2
ecdsa==0.10
gunicorn==18.0
paramiko==1.12.0
pycrypto==2.6.1
six==1.4.1
sorl-thumbnail==11.12
wsgiref==0.1.2
